test: move mirror() to test/fake_externals
